requirements.txt
opencv-python
numpy
pillow
pyserial
tensorflow
requests
python-dotenv
serial
